# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.Lote:
    properties:
      data_fornecimento:
        type: string
      estragados:
        type: integer
      id_fornecedor:
        type: integer
      id_lote:
        type: integer
      id_produto:
        type: integer
      preco_unitario:
        type: number
      quantidade_inicial:
        type: integer
      validade:
        type: string
    type: object
  model.LoteCreate:
    properties:
      data_fornecimento:
        type: string
      estragados:
        type: integer
      id_fornecedor:
        type: integer
      id_produto:
        type: integer
      preco_unitario:
        type: number
      quantidade_inicial:
        type: integer
      validade:
        type: string
    type: object
...
  services_lote.GastoMensal:
    properties:
      lotes_comprados:
        type: integer
      mes:
        type: integer
      total_gasto:
        type: number
    type: object
  types.ErrorResponse:
    properties:
      detail:
        type: string
    type: object
info:
  contact: {}
  description: Aplicação de Banco de Dados para Gerenciamento de Bares
  title: EDNA Bar System
  version: "1.0"
paths:
...
  /lotes:
    get:
      parameters:
      - description: 'Filter by nome using operators: like, ilike, eq, ne. Format:
          operator.value (e.g. like.João)'
        in: query
        name: filter-nome
        type: string
      - description: 'Filter by cnpj using operators: eq, ne, like, ilike. Format:
          operator.value (e.g. eq.123456789)'
        in: query
        name: filter-cnpj
        type: string
      - description: 'Sort fields: nome, cnpj. Prefix with ''-'' for desc. Comma separated
          for multiple fields (e.g. -nome,cnpj)'
        in: query
        name: sort
        type: string
      - description: Pagination offset (default 0)
        in: query
        name: offset
        type: integer
      - description: Pagination limit (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Lote'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: List Lotes
      tags:
      - Lote
    post:
      consumes:
      - application/json
      parameters:
      - description: Lote payload
        in: body
        name: fornecedor
        required: true
        schema:
          $ref: '#/definitions/model.LoteCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Lote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Create Lote
      tags:
      - Lote
  /lotes/{id}:
    delete:
      parameters:
      - description: Lote ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Lote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Delete Lote
      tags:
      - Lote
    get:
      parameters:
      - description: Lote ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Lote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get Lote by ID
      tags:
      - Lote
    put:
      consumes:
      - application/json
      parameters:
      - description: Lote ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lote payload
        in: body
        name: fornecedor
        required: true
        schema:
          $ref: '#/definitions/model.LoteCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Lote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Update Lote
      tags:
      - Lote
  /lotes/relatorio:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/services_lote.GastoMensal'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Get Relatorio of Lotes spends
      tags:
      - Lote
...
swagger: "2.0"